use custom hook for swiping
